+++ Sources/SourceDocs/Commands/Clean.swift
@@ -16,7 +16,8 @@ struct CleanCommandOptions: OptionsProtocol {
 
     static func evaluate(_ mode: CommandMode) -> Result<CleanCommandOptions, CommandantError<SourceDocsError>> {
         return curry(self.init)
-            <*> mode <| Option(key: "output-folder", defaultValue: SourceDocs.defaultOutputPath, usage: "Output directory (defaults to Docs/Reference).")
+            <*> mode <| Option(key: "output-folder", defaultValue: SourceDocs.defaultOutputPath,
+                               usage: "Output directory (defaults to \(SourceDocs.defaultOutputPath)).")
     }
 }